+++ integration-tests/src/test/java/io/nem/symbol/sdk/infrastructure/VotingKeyLinkTransactionIntegrationTest.java
@@ -16,44 +16,16 @@
 package io.nem.symbol.sdk.infrastructure;
 
 import io.nem.symbol.core.crypto.PublicKey;
-import io.nem.symbol.core.crypto.VotingKey;
 import io.nem.symbol.sdk.model.account.Account;
 import io.nem.symbol.sdk.model.transaction.LinkAction;
 import io.nem.symbol.sdk.model.transaction.VotingKeyLinkTransaction;
 import io.nem.symbol.sdk.model.transaction.VotingKeyLinkTransactionFactory;
-import io.nem.symbol.sdk.model.transaction.VotingKeyLinkV1Transaction;
-import io.nem.symbol.sdk.model.transaction.VotingKeyLinkV1TransactionFactory;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class VotingKeyLinkTransactionIntegrationTest extends BaseIntegrationTest {
-
-  @ParameterizedTest
-  @EnumSource(RepositoryType.class)
-  public void basicAnnounceV1(RepositoryType type) {
-
-    Account account = config().getNemesisAccount2();
-
-    VotingKey linkedPublicKey = VotingKey.generateRandom();
-
-    VotingKeyLinkV1Transaction linkTransaction =
-        VotingKeyLinkV1TransactionFactory.create(
-                getNetworkType(), getDeadline(), linkedPublicKey, (72), (26280), LinkAction.LINK)
-            .maxFee(maxFee)
-            .build();
-
-    announceAndValidate(type, account, linkTransaction);
-
-    VotingKeyLinkV1Transaction unlinkTransaction =
-        VotingKeyLinkV1TransactionFactory.create(
-                getNetworkType(), getDeadline(), linkedPublicKey, (72), (26280), LinkAction.UNLINK)
-            .maxFee(maxFee)
-            .build();
-
-    announceAndValidate(type, account, unlinkTransaction);
-  }
 
   @ParameterizedTest
   @EnumSource(RepositoryType.class)