+++ sdk-core/src/test/java/io/nem/symbol/sdk/model/account/AddressVectorTester.java
@@ -38,7 +38,7 @@ class AddressVectorTester extends AbstractVectorTester {
     arguments.add(extractArguments(entry, NetworkType.MAIN_NET, "address_Public"));
     arguments.add(extractArguments(entry, NetworkType.TEST_NET, "address_PublicTest"));
     arguments.add(extractArguments(entry, NetworkType.PRIVATE, "address_Private"));
-    arguments.add(extractArguments(entry, NetworkType.PRIVATE_TEST, "address_PrivateTest"));
+    //    arguments.add(extractArguments(entry, NetworkType.PRIVATE_TEST, "address_PrivateTest"));
     return arguments;
   }