+++ sdk-vertx-client/src/main/java/io/nem/symbol/sdk/infrastructure/vertx/NetworkRepositoryVertxImpl.java
@@ -98,8 +98,11 @@ public class NetworkRepositoryVertxImpl extends AbstractRepositoryVertxImpl
             .map(
                 info ->
                     new TransactionFees(
-                        info.getAverageFeeMultiplier(), info.getMedianFeeMultiplier(),
-                        info.getLowestFeeMultiplier(), info.getHighestFeeMultiplier())));
+                        info.getAverageFeeMultiplier(),
+                        info.getMedianFeeMultiplier(),
+                        info.getLowestFeeMultiplier(),
+                        info.getHighestFeeMultiplier(),
+                        info.getMinFeeMultiplier())));
   }
 
   @Override