+++ sdk-vertx-client/src/main/java/io/nem/symbol/sdk/infrastructure/vertx/mappers/GeneralTransactionMapper.java
@@ -62,7 +62,6 @@ public class GeneralTransactionMapper implements TransactionMapper {
     register(new AccountOperationRestrictionTransactionMapper(jsonHelper));
     register(new NodeKeyLinkTransactionMapper(jsonHelper));
     register(new VotingKeyLinkTransactionMapper(jsonHelper));
-    register(new VotingKeyLinkV1TransactionMapper(jsonHelper));
     register(new VrfKeyLinkTransactionMapper(jsonHelper));
     register(new AccountKeyLinkTransactionMapper(jsonHelper));
     register(new AggregateTransactionMapper(jsonHelper, TransactionType.AGGREGATE_BONDED, this));