+++ sdk-vertx-client/src/test/java/io/nem/symbol/sdk/infrastructure/vertx/FinalizationRepositoryVertxImplTest.java
@@ -76,7 +76,6 @@ public class FinalizationRepositoryVertxImplTest extends AbstractVertxRespositor
     dto.height(BigInteger.valueOf(4));
 
     MessageGroup messageGroup = new MessageGroup();
-    messageGroup.setSignatureSchema(10);
     messageGroup.stage(StageEnum.NUMBER_1);
     messageGroup.addHashesItem("hash1");
     messageGroup.setHeight(BigInteger.valueOf(20));