+++ CHANGELOG.md
@@ -3,6 +3,24 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.23.0] - 14-Jan-2021
+
+**Milestone**: Catapult-server finality (0.10.0.4)
+ Package  | Version  | Link
+---|---|---
+SDK OkHttp | v0.23.0 | https://repo.maven.apache.org/maven2/io/nem/symbol-sdk-okhttp-client
+SDK Vertx | v0.23.0 | https://repo.maven.apache.org/maven2/io/nem/symbol-sdk-vertx-client
+Catbuffer Library | v0.1.1 | https://repo.maven.apache.org/maven2/io/nem/catbuffer-java
+Client OkHttp | v0.11.1  | https://repo.maven.apache.org/maven2/io/nem/symbol-openapi-okhttp-gson-client
+Client Vertx | v0.11.1  | https://repo.maven.apache.org/maven2/io/nem/symbol-openapi-vertx-client/
+
+- **Re track to catapult-server main branch**
+- **All entity versions reset to v1 for Symbol mainnet release**
+- **[BREAKING CHANGE]** Finalization proof middle removed `messageGroup.schemaVersion` and `BMTreeSignaure.top`
+- **[BREAKING CHANGE]** Removed `VotingKeyLinkV1Transaction`.
+- **[BREAKING CHANGE]** `PrivateTest` network type changed from `0x80` to `0xA8`.
+- Added `minFeeMultiplier` in transaction fee endpoint.
+
 ## [0.22.1] - 14-Dec-2020
 
 **Milestone**: Catapult-server finality (0.10.0.4)